// Front-End/src/app/models/vehicle.entry.checklist.ts
export class VehicleEntryChecklist {
    companyId: number | null = null;
    resaleId: number | null = null;
    id: number | null = null;

    checklist1Enabled: string | null = null;
    checklist1Label: string = '';
    checklist1Desc: string = '';

    checklist2Enabled: string | null = null;
    checklist2Label: string = '';
    checklist2Desc: string = '';

    checklist3Enabled: string | null = null;
    checklist3Label: string = '';
    checklist3Desc: string = '';

    checklist4Enabled: string | null = null;
    checklist4Label: string = '';
    checklist4Desc: string = '';

    checklist5Enabled: string | null = null;
    checklist5Label: string = '';
    checklist5Desc: string = '';

    checklist6Enabled: string | null = null;
    checklist6Label: string = '';
    checklist6Desc: string = '';

    checklist7Enabled: string | null = null;
    checklist7Label: string = '';
    checklist7Desc: string = '';

    checklist8Enabled: string | null = null;
    checklist8Label: string = '';
    checklist8Desc: string = '';

    checklist9Enabled: string | null = null;
    checklist9Label: string = '';
    checklist9Desc: string = '';

    checklist10Enabled: string | null = null;
    checklist10Label: string = '';
    checklist10Desc: string = '';

    checklist11Enabled: string | null = null;
    checklist11Label: string = '';
    checklist11Desc: string = '';

    checklist12Enabled: string | null = null;
    checklist12Label: string = '';
    checklist12Desc: string = '';

    checklist13Enabled: string | null = null;
    checklist13Label: string = '';
    checklist13Desc: string = '';

    checklist14Enabled: string | null = null;
    checklist14Label: string = '';
    checklist14Desc: string = '';

    checklist15Enabled: string | null = null;
    checklist15Label: string = '';
    checklist15Desc: string = '';

    checklist16Enabled: string | null = null;
    checklist16Label: string = '';
    checklist16Desc: string = '';

    checklist17Enabled: string | null = null;
    checklist17Label: string = '';
    checklist17Desc: string = '';

    checklist18Enabled: string | null = null;
    checklist18Label: string = '';
    checklist18Desc: string = '';

    checklist19Enabled: string | null = null;
    checklist19Label: string = '';
    checklist19Desc: string = '';

    checklist20Enabled: string | null = null;
    checklist20Label: string = '';
    checklist20Desc: string = '';
}